@@ -1,14 +1,14 @@
-﻿namespace UtilityLibraries
-{
-    public static class StringLibrary
-    {
-        public static bool StartsWithUpper(this string str)
-        {
-            if (string.IsNullOrWhiteSpace(str))
-                return false;
+﻿namespace UtilityLibraries;
 
-            char ch = str[0];
-            return char.IsUpper(ch);
-        }
+public static class StringLibrary
+{
+    public static bool StartsWithUpper(this string? str)
+    {
+        if (string.IsNullOrWhiteSpace(str))
+            return false;
+
+        char ch = str[0];
+        return char.IsUpper(ch);
     }
 }
+
