@@ -1,52 +1,52 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using UtilityLibraries;
 
-namespace StringLibraryTest
+namespace StringLibraryTest;
+
+[TestClass]
+public class UnitTest1
 {
-    [TestClass]
-    public class UnitTest1
+    [TestMethod]
+    public void TestStartsWithUpper()
     {
-        [TestMethod]
-        public void TestStartsWithUpper()
+        // Tests that we expect to return true.
+        string[] words = { "Alphabet", "Zebra", "ABC", "Αθήνα", "Москва" };
+        foreach (var word in words)
         {
-            // Tests that we expect to return true.
-            string[] words = { "Alphabet", "Zebra", "ABC", "Αθήνα", "Москва" };
-            foreach (var word in words)
-            {
-                bool result = word.StartsWithUpper();
-                Assert.IsTrue(result,
-                       String.Format("Expected for '{0}': true; Actual: {1}",
-                                     word, result));
-            }
+            bool result = word.StartsWithUpper();
+            Assert.IsTrue(result,
+                   String.Format("Expected for '{0}': true; Actual: {1}",
+                                 word, result));
         }
+    }
 
-        [TestMethod]
-        public void TestDoesNotStartWithUpper()
-        {
-            // Tests that we expect to return false.
-            string[] words = { "alphabet", "zebra", "abc", "αυτοκινητοβιομηχανία", "государство",
+    [TestMethod]
+    public void TestDoesNotStartWithUpper()
+    {
+        // Tests that we expect to return false.
+        string[] words = { "alphabet", "zebra", "abc", "αυτοκινητοβιομηχανία", "государство",
                                "1234", ".", ";", " " };
-            foreach (var word in words)
-            {
-                bool result = word.StartsWithUpper();
-                Assert.IsFalse(result,
-                       String.Format("Expected for '{0}': false; Actual: {1}",
-                                     word, result));
-            }
-        }
-
-        [TestMethod]
-        public void DirectCallWithNullOrEmpty()
+        foreach (var word in words)
         {
-            // Tests that we expect to return false.
-            string[] words = { string.Empty, null };
-            foreach (var word in words)
-            {
-                bool result = StringLibrary.StartsWithUpper(word);
-                Assert.IsFalse(result,
-                       String.Format("Expected for '{0}': false; Actual: {1}",
-                                     word == null ? "<null>" : word, result));
-            }
+            bool result = word.StartsWithUpper();
+            Assert.IsFalse(result,
+                   String.Format("Expected for '{0}': false; Actual: {1}",
+                                 word, result));
+        }
+    }
+
+    [TestMethod]
+    public void DirectCallWithNullOrEmpty()
+    {
+        // Tests that we expect to return false.
+        string?[] words = { string.Empty, null };
+        foreach (var word in words)
+        {
+            bool result = StringLibrary.StartsWithUpper(word);
+            Assert.IsFalse(result,
+                   String.Format("Expected for '{0}': false; Actual: {1}",
+                                 word == null ? "<null>" : word, result));
         }
     }
 }
+
