@@ -33,28 +33,25 @@ namespace SerializeOnlyNoOptions
     {
         public static void Main()
         {
+            string jsonString;
             WeatherForecast weatherForecast = new()
                 { Date = DateTime.Parse("2019-08-01"), TemperatureCelsius = 25, Summary = "Hot" };
 
-            // Use a context that selects Serialization mode.
-            // <SerializeOnlyContext> 
-            using MemoryStream memoryStream1 = new();
-            using Utf8JsonWriter writer1 = new(memoryStream1);
-            SerializeOnlyContext.Default.WeatherForecast.Serialize(writer1, weatherForecast);
-            writer1.Flush();
-            // </SerializeOnlyContext> 
-            Console.WriteLine(Encoding.UTF8.GetString(memoryStream1.ToArray()));
+            // Use context that selects Serialization mode only for WeatherForecast.
+            // <SerializeWFContext>
+            jsonString = JsonSerializer.Serialize(weatherForecast,
+                SerializeOnlyWeatherForecastOnlyContext.Default.WeatherForecast);
+            // </SerializeWFContext> 
+            Console.WriteLine(jsonString);
             // output:
             //{"Date":"2019-08-01T00:00:00","TemperatureCelsius":25,"Summary":"Hot"}
 
-            // Use context that selects Serialization mode only for WeatherForecast.
-            // <SerializeWFContext> 
-            using MemoryStream memoryStream2 = new();
-            using Utf8JsonWriter writer2 = new(memoryStream2);
-            SerializeOnlyContext.Default.WeatherForecast.Serialize(writer2, weatherForecast);
-            writer2.Flush();
-            // </SerializeWFContext> 
-            Console.WriteLine(Encoding.UTF8.GetString(memoryStream2.ToArray()));
+            // Use a context that selects Serialization mode.
+            // <SerializeOnlyContext> 
+            jsonString = JsonSerializer.Serialize(weatherForecast,
+                SerializeOnlyContext.Default.WeatherForecast);
+            // </SerializeOnlyContext> 
+            Console.WriteLine(jsonString);
             // output:
             //{"Date":"2019-08-01T00:00:00","TemperatureCelsius":25,"Summary":"Hot"}
         }
