@@ -33,20 +33,6 @@ namespace SerializeOnlyWithOptions
             WeatherForecast weatherForecast = new()
                 { Date = DateTime.Parse("2019-08-01"), TemperatureCelsius = 25, Summary = "Hot" };
 
-            // Serialize using Default context
-            // and options specified by [JsonSourceGenerationOptions].
-            // <SerializeWithContext>
-            jsonString = JsonSerializer.Serialize(
-                weatherForecast, typeof(WeatherForecast), SerializationModeOptionsContext.Default);
-            // </SerializeWithContext>
-            Console.WriteLine(jsonString);
-            // output:
-            //{
-            //  "date": "2019-08-01T00:00:00",
-            //  "temperatureCelsius": 0,
-            //  "summary": "Hot"
-            //}
-
             // Serialize using TypeInfo<TValue> provided by the context
             // and options specified by [JsonSourceGenerationOptions].
             // <SerializeWithTypeInfo>
@@ -61,16 +47,19 @@ namespace SerializeOnlyWithOptions
             //  "summary": "Hot"
             //}
 
-            // <SerializeDirect>
-            using MemoryStream stream = new();
-            using Utf8JsonWriter writer = new(stream);
-            SerializationModeOptionsContext.Default.WeatherForecast.Serialize(
-                writer, weatherForecast);
-            writer.Flush();
-            // </SerializeDirect>
-            Console.WriteLine(Encoding.UTF8.GetString(stream.ToArray()));
+            // Serialize using Default context
+            // and options specified by [JsonSourceGenerationOptions].
+            // <SerializeWithContext>
+            jsonString = JsonSerializer.Serialize(
+                weatherForecast, typeof(WeatherForecast), SerializationModeOptionsContext.Default);
+            // </SerializeWithContext>
+            Console.WriteLine(jsonString);
             // output:
-            //{"date":"2019-08-01T00:00:00","temperatureCelsius":25,"summary":"Hot"}
+            //{
+            //  "date": "2019-08-01T00:00:00",
+            //  "temperatureCelsius": 0,
+            //  "summary": "Hot"
+            //}
         }
     }
 }
