@@ -4,7 +4,7 @@ let months = [| "January"; "February"; "March"; "April";
 
 let lookupMonth month =
    if (month > 12 || month < 1)
-     then invalidArg "month" (sprintf "Value passed in was %d." month)
+     then invalidArg (nameof month) (sprintf "Value passed in was %d." month)
    months.[month - 1]
 
 printfn "%s" (lookupMonth 12)
